+++ wsidicom/web/wsidicom_web_image_data.py
@@ -19,6 +19,7 @@ from PIL import Image
 from PIL.Image import Image as PILImage
 from pydicom.uid import UID
 
+from wsidicom.codec import Codec
 from wsidicom.geometry import Point
 from wsidicom.instance import WsiDataset, WsiDicomImageData
 from wsidicom.web.wsidicom_web_client import WsiDicomWebClient
@@ -55,7 +56,14 @@ class WsiDicomWebImageData(WsiDicomImageData):
         self._series_uid = dataset.uids.slide.series_instance
         self._instance_uid = dataset.uids.instance
         self._transfer_syntax = transfer_syntax
-        super().__init__([dataset])
+        codec = Codec.create(
+            self.transfer_syntax,
+            dataset.samples_per_pixel,
+            dataset.bits,
+            dataset.tile_size,
+            dataset.photometric_interpretation,
+        )
+        super().__init__([dataset], codec)
 
     @property
     def transfer_syntax(self) -> UID: